INTERLOK-3781 Repackaging away from com.adaptris.rest to have sub-packages for each project.  A few changes to access modifiers have been needed.
