+++ interlok-rest-cluster/src/main/resources/META-INF/com/adaptris/core/management/components/cluster-rest
@@ -1,1 +1,1 @@
-class=com.adaptris.rest.ClusterManagerComponent
+class=com.adaptris.rest.cluster.ClusterManagerComponent
