+++ interlok-rest-health-check/src/main/resources/META-INF/com/adaptris/core/management/components/health-check
@@ -1,1 +1,1 @@
-class=com.adaptris.rest.WorkflowHealthCheckComponent
+class=com.adaptris.rest.healthcheck.WorkflowHealthCheckComponent
