+++ interlok-rest-provider-datadog/src/main/java/com/adaptris/rest/metrics/datadog/DatadogPushComponent.java
@@ -1,4 +1,4 @@
-package com.adaptris.rest;
+package com.adaptris.rest.metrics.datadog;
 
 import java.time.Duration;
 import java.util.Properties;
