+++ interlok-rest-provider-datadog/src/main/resources/META-INF/com/adaptris/core/management/components/metrics-provider-datadog
@@ -1,1 +1,1 @@
-class=com.adaptris.rest.DatadogPushComponent
+class=com.adaptris.rest.metrics.datadog.DatadogPushComponent
