+++ interlok-rest-provider-prometheus/src/main/resources/META-INF/com/adaptris/core/management/components/metrics-provider-prometheus
@@ -1,1 +1,1 @@
-class=com.adaptris.rest.PrometheusEndpointComponent
+class=com.adaptris.rest.metrics.prometheus.PrometheusEndpointComponent
